增加插入专用格

+++ common/src/main/java/io/yukkuric/hexop/hexal/MoteChestContainer.java
@@ -7,10 +7,16 @@ import net.minecraft.world.item.ItemStack;
 import static io.yukkuric.hexop.hexal.CompressedChestMenu.UI_VISIBLE_SLOTS;
 
 public class MoteChestContainer implements Container {
-    Container parent;
+    final Container parent;
+    final CachedNexusInventory API;
 
-    public MoteChestContainer(Container wrapper) {
+    public MoteChestContainer(Container wrapper, CachedNexusInventory api) {
         parent = wrapper;
+        API = api;
+    }
+
+    boolean shouldKeepLastCellEmpty(int i) {
+        return i == UI_VISIBLE_SLOTS - 1 && !API.isFull();
     }
 
     public int getContainerSize() {
@@ -24,21 +30,28 @@ public class MoteChestContainer implements Container {
 
     @Override
     public ItemStack getItem(int i) {
+        if (shouldKeepLastCellEmpty(i)) return ItemStack.EMPTY;
         return parent.getItem(i);
     }
 
     @Override
     public ItemStack removeItem(int i, int j) {
+        if (shouldKeepLastCellEmpty(i)) return ItemStack.EMPTY;
         return parent.removeItem(i, j);
     }
 
     @Override
     public ItemStack removeItemNoUpdate(int i) {
+        if (shouldKeepLastCellEmpty(i)) return ItemStack.EMPTY;
         return parent.removeItemNoUpdate(i);
     }
 
     @Override
     public void setItem(int i, ItemStack itemStack) {
+        if (shouldKeepLastCellEmpty(i)) {
+            API.insertItem(i, itemStack, false);
+            return;
+        }
         parent.setItem(i, itemStack);
     }
